feat(input): add discrete event pipeline with evdev keyboard source and audio toggle
Introduce an InputEvent type hierarchy (Key, MidiNote, MidiControlChange,
MidiPitchBend, MidiRealtime) in hypercolor-types to carry discrete host
input events alongside the existing sampled InteractionInput data. Add
InputButtonState (Pressed/Released/Repeated) and MidiRealtimeMessage
enums, and wire InputEventReceived into the event bus.

Implement EvdevKeyboardInput as the first concrete event source on Linux.
A background worker thread polls /dev/input/event* devices at 8 ms
intervals, filters for keyboards by capability, and maps 70+ evdev key
codes to canonical names. The InputSource trait gains a drain_events()
method with a default no-op so sampled-only sources remain unchanged.

Improve audio capture lifecycle by parking the previous AudioSourceType
when live audio is disabled instead of tearing down the stream. Resuming
the same source skips reconstruction entirely. A new effective-capture
flag prevents stream creation when the source is set to None.

Add a SidebarAudioToggle component to the Leptos UI that shows a
waveform icon with a beat-pulse glow when audio is active and a muted
icon when disabled. The toggle calls the daemon config API and the
audio_enabled signal is seeded from daemon state on startup.

Add tests for event draining across multiple sources, render-thread
event publication, audio enable/disable round-trips, and InputEvent
JSON serialization. Fix redundant casts in canvas bilinear interpolation
and backtick-wrap LightScript in effect docs.

+++ crates/hypercolor-types/src/canvas.rs
@@ -637,9 +637,9 @@ impl Canvas {
         let x1 = (x0 + 1).min(self.width - 1);
         let y1 = (y0 + 1).min(self.height - 1);
 
-        let frac_x = (fx.fract() * BILINEAR_ONE as f32).clamp(0.0, BILINEAR_ONE as f32) as f32
+        let frac_x = (fx.fract() * BILINEAR_ONE as f32).clamp(0.0, BILINEAR_ONE as f32)
             / BILINEAR_ONE as f32;
-        let frac_y = (fy.fract() * BILINEAR_ONE as f32).clamp(0.0, BILINEAR_ONE as f32) as f32
+        let frac_y = (fy.fract() * BILINEAR_ONE as f32).clamp(0.0, BILINEAR_ONE as f32)
             / BILINEAR_ONE as f32;
 
         let top = RgbaF32::lerp(
